backend/requirements.txt
fastapi==0.132.0
uvicorn==0.41.0
google-genai==1.64.0
PyMuPDF==1.27.1
pydantic==2.12.5
python-dotenv==1.2.1
numpy==2.4.2
scikit-learn==1.8.0
